split gt providers
