+++ packages/react/src/context.types.ts
@@ -1,9 +1,9 @@
-import type { GTProvider as RuntimeGTProvider } from './provider/GTProvider';
+import type { ServerGTProvider } from './provider/ServerGTProvider';
 
 /**
  * Wrap GTProvider around the content that you want to translate
  */
-export const GTProvider: typeof RuntimeGTProvider = () => {
+export const GTProvider: typeof ServerGTProvider = () => {
   throw new Error(
     'gt-react: You have imported a function from the dedicated types entrypoint. If you are seeing this error, it means something has gone wrong.'
   );
